pyproject.toml
[tool.poetry]
name = "pyiconlibrary"
version = "1.0.0"
description = "A class that uses icons.yaml to allow a user to load any icon in the google material icons library. To pick an icon, simply find the icon you want on https://fonts.google.com/icons"
authors = ["Philip Mai <philip.mai@accenture.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.8"
black= "^22.12.0"
build= "^0.9.0"
click= "^8.1.3"
customtkinter= "^5.0.3"
darkdetect= "^0.8.0"
flake8= "^5.0.4"
isort= "^5.11.4"
mccabe= "^0.7.0"
mypy-extensions= "^0.4.3"
packaging= "^22.0"
pathspec= "^0.10.3"
pep517= "^0.13.0"
Pillow= "^9.3.0"
platformdirs= "^2.6.2"
pycodestyle= "^2.9.1"
pyflakes= "^2.5.0"
pyproject-flake8= "^5.0.4"
PyYAML= "^6.0"
tomli= "^2.0.1"
typing-extensions= "^4.4.0"
Wand= "^0.6.10"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 101

[tool.flake8]
max-line-length = 101
extend-ignore = ['402','F841','F401','E302','E305','E203','W503','C901','W1514', 'F403']
exclude = ['.git','__pycache__','build','devel','logs','.catkin_tools']
max-complexity = 10

[tool.isosort]
profile = "black"

[tool.pylint.message_control]
disable = ["import-error"]
